Check if backups dir is empty before removing.
This is a workaround for the unexpected behavior of rmdir in the 
emulator yuzu. On console, rmdir would only remove the directory if it 
is empty. This is the behavior I intended and expected. But in yuzu 
rmdir will remove the directory and all of its contents, causing the all 
of the remaining backups to be removed after the first is restored. By 
explicitly checking if it is empty first, this will work as intended on 
console and in yuzu.

+++ source/mod_installer.h
@@ -360,7 +360,10 @@ void load_mods(FILE* f_arc) {
                 }
                 std::string parentPath = std::filesystem::path(backup_path).parent_path();
                 delete[] backup_path;
-                rmdir(parentPath.c_str());
+                std::error_code ec;
+                if(std::filesystem::is_empty(parentPath, ec)) {
+                    rmdir(parentPath.c_str());
+                }
             }
         }
     }
